backend/requirements.txt
asgiref==1.1.2
attrs==19.3.0
autobahn==20.2.1
Automat==0.8.0
certifi==2019.11.28
cffi==1.14.0
channels==1.1.8
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
daphne==1.4.2
Django==2.2
django-cors-headers==3.1.1
django-filter==2.2.0
django-oauth-toolkit==1.2.0
djangorestframework==3.11.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
oauthlib==3.1.0
psycopg2==2.8.3
pycparser==2.19
PyHamcrest==2.0.0
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
Twisted==19.10.0
txaio==20.1.1
urllib3==1.25.8
zope.interface==4.7.1
